Do not panic on NaN
The datadog client insists on float64s, which were just assumed to be
always present in the JSON data from presto. However if presto was
starting up then some numbers were returned as NaN. NaN is not a valid
number in JSON so are encoded as the string, "NaN". When they were
decoded on our side, this converted them to a string, not a float64 and
the blind conversion in sendJMXMetric would fail.

This change comes with no guarentee. If it breaks your application, you
get to keep both pieces.

Fixes presto#1